Another cache, provides noticeable performance increase and reduces memory usage:

Before: 
80 MB   2.72s
107 MB  3.82s
159 MB  8.27s
After: 
80 MB   2.18s
107 MB  2.99s
159 MB  6.49s

Memory before: 4754.88MB 
Memory after:   3495.16MB
(for one demo file - 159 MB)